deployment/requirements.txt
altair
attrs
blinker
cachetools
certifi
charset-normalizer
click
colorama
contourpy
cycler
feature-engine
fonttools
gitdb
GitPython
idna
importlib-metadata
Jinja2
joblib
jsonschema
jsonschema-specifications
kiwisolver
markdown-it-py
MarkupSafe
matplotlib
mdurl
numpy
packaging
pandas
patsy
phik
Pillow
protobuf
pyarrow
pydeck
Pygments
pyparsing
python-dateutil
pytz
referencing
requests
rich
rpds-py
scikit-learn
scipy
seaborn
six
smmap
statsmodels
streamlit
tenacity
threadpoolctl
toml
toolz
tornado
typing_extensions
tzdata
tzlocal
urllib3
validators
watchdog
zipp